refactor(renderer): move pocket, creatables, asset-tags to assets/
Step 4 of the renderer reorganization. The surfaces that move Weave assets
around — the pocket, creatables, and asset tags — gather under a new
top-level assets/.

Pure file moves + import-path updates; no behaviour change.
