+++ src/renderer/src/assets/creatables/creatable-view.ts
