+++ src/renderer/src/assets/pocket/search-result-element.ts
